requirements.txt
python-dotenv
Office365-REST-Python-Client
